reworked supabase folder
reorganized supabase package,
added functions for querying/updating bounties
fixed issues preventing the agent from correctly loading the package
