+++ packages/plugin-vsa-contracts/package.json
@@ -25,6 +25,7 @@
         "@elizaos/core": "workspace:*",
         "@ethersproject/abi": "^5.7.0",
         "@elizaos/nillion-core": "workspace:*",
+        "@elizaos/supabase": "workspace:*",
         "@ethersproject/abstract-provider": "^5.7.0",
         "@ethersproject/address": "^5.7.0",
         "@ethersproject/bignumber": "^5.7.0",
