+++ packages/plugin-vsa-contracts/src/actions/createUserProfileAction.ts
@@ -135,7 +135,7 @@ export const createUserProfileAction: Action = {
                 modelClass: ModelClass.LARGE,
             }));
     
-            console.log("content stage mate");
+            console.log("content stage");
             console.log(content);
     
             if (!isUserData(content)) {
